Added Main Menu scene back to the project.
Ok, the Main Menu has been brought back to the project. I made sure to include all its scripts in the Scripts folder. I also added an Options button that looks identical to the Start Game button. I'll leave the Options_Button script to Ryan, and that's everything so far.